feat: profile picture component
